Changes to shorten
I removed the pieces of the article which show the 'warts' of Azure PowerShell and the fact it's not all one thing yet.  Because the links to the different PowerShell areas are on the left nav I don't want to highlight the fact that we're not completely organized yet.  It's hard to explain why to the customer.